+++ src/main/java/http/controller/Controller.java
@@ -1,7 +1,7 @@
 package http.controller;
 
-import http.HttpRequest;
-import http.HttpResponse;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 import utils.HttpResponseHeaderParser;
 
 public abstract class Controller {
