+++ src/main/java/http/request/HeaderParam.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 public enum HeaderParam {
     HOST("host"),
