+++ src/main/java/http/request/HttpRequest.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
