+++ src/main/java/http/request/RequestBody.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 import utils.IOUtils;
