+++ src/main/java/webserver/RequestHandler.java
@@ -1,11 +1,11 @@
 package webserver;
 
-import http.HttpRequest;
-import http.HttpResponse;
-import http.RequestMethod;
 import http.controller.Controller;
 import http.controller.Controllers;
 import http.controller.ControllersFactory;
+import http.request.HttpRequest;
+import http.request.RequestMethod;
+import http.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.HttpResponseHeaderParser;
