+++ src/test/java/service/UserServiceTest.java
@@ -1,7 +1,7 @@
 package service;
 
 import db.DataBase;
-import http.HttpRequest;
+import http.request.HttpRequest;
 import model.User;
 import org.junit.jupiter.api.Test;
